@@ -7,7 +7,7 @@
     "license": "MIT",
     "require": {
         "php": "^8.2",
-        "laravel/framework": "^11.31",
+        "laravel/framework": "^v12.0.1",
         "laravel/sanctum": "^4.0",
         "laravel/tinker": "^2.9"
     },
