@@ -2,21 +2,22 @@
 
 namespace App\Mail\User\Auth;
 
+use App\Models\User;
 use Illuminate\Mail\Mailable;
 
 class SuccessRegistrationEmail extends Mailable
 {
-    public $user;
+    public User $user;
 
     public function __construct($user)
     {
         $this->user = $user;
     }
 
-    public function build()
+    public function build(): SuccessRegistrationEmail
     {
         return $this->to($this->user->email)
-        ->subject('Добро пожаловать в наш сервис!')
+            ->subject('Добро пожаловать в наш сервис!')
             ->view('mail.auth.SuccessRegistrationNotification')
             ->with([
                 'name' => $this->user->name,
