@@ -10,7 +10,8 @@ class ProductIndexAction
 {
     public function __construct(
         protected ProductIndexService $service,
-    ) {}
+    ) {
+    }
 
     /**
      * @throws ProductNotFound
