@@ -9,7 +9,8 @@ class UserForgotPasswordEmailFormAction
 {
     public function __construct(
         protected UserForgotPasswordEmailFormService $service
-    ) {}
+    ) {
+    }
 
     /**
      * @throws UserForgotPasswordEmailFormException
