@@ -10,7 +10,8 @@ class UserLoginAction
 {
     public function __construct(
         protected UserLoginService $service,
-    ) {}
+    ) {
+    }
 
     /**
      * @throws UserLoginException
