@@ -10,7 +10,8 @@ class UserRegisterAction
 {
     public function __construct(
         protected UserRegisterService $service
-    ) {}
+    ) {
+    }
 
     /**
      * @throws UserRegisterException
