@@ -9,7 +9,8 @@ class UserResetPasswordFormAction
 {
     public function __construct(
         protected UserResetPasswordFormService $service
-    ) {}
+    ) {
+    }
 
     /**
      * @throws UserResetPasswordFormException
