@@ -14,7 +14,8 @@ class UserLoginService
 {
     public function __construct(
         protected UserRepository $repository,
-    ) {}
+    ) {
+    }
 
     /**
      * @throws UserLoginException|Throwable
