@@ -14,7 +14,8 @@ class UserRegisterService
 {
     public function __construct(
         protected UserRepository $repository
-    ) {}
+    ) {
+    }
 
     /**
      * @throws UserRegisterException
