@@ -2,11 +2,9 @@
 
 use App\Http\Controllers\MainController;
 use App\Http\Controllers\ProductController;
-use App\Http\Controllers\User\Auth\ForgotPasswordController;
 use App\Http\Controllers\User\Auth\LoginController;
-use App\Http\Controllers\User\Auth\RegisterController;
-use App\Http\Controllers\User\Auth\ResetPasswordController;
 use App\Http\Controllers\User\Auth\VerificationController;
+use App\Http\Controllers\User\UserController;
 use Illuminate\Support\Facades\Route;
 
 Route::get('/', [MainController::class, 'index'])
@@ -15,31 +13,14 @@ Route::get('/', [MainController::class, 'index'])
 Route::post('locale', [MainController::class, 'locale'])
     ->name('locale.change');
 
-Route::middleware('guest')->group(function () {
-
-    Route::get('register', [RegisterController::class, 'index'])
-        ->name('register');
-    Route::post('register', [RegisterController::class, 'store'])
-        ->name('register.store');
-
-    Route::get('login', [LoginController::class, 'index'])
-        ->name('login');
-    Route::post('login', [LoginController::class, 'store'])
-        ->name('login.store');
-
-    Route::get('password/reset', [ForgotPasswordController::class, 'index'])
-        ->name('password.request');
-    Route::post('password/email', [ForgotPasswordController::class, 'store'])
-        ->name('password.email');
-
-    Route::get('password/reset/{token}', [ResetPasswordController::class, 'index'])
-        ->name('password.reset');
-    Route::post('password/reset', [ResetPasswordController::class, 'store'])
-        ->name('password.update');
-
-});
-
 Route::middleware('auth')->group(function () {
+
+    Route::get('/profile/{user_id}', [UserController::class, 'index'])->name('profile');
+    Route::put('/profile/update', [UserController::class, 'store'])->name('profile.update');
+    Route::post('/profile/add/address', [UserController::class, 'addAddress'])->name('profile.add.address');
+    Route::put('/profile/password/update', [UserController::class, 'password'])->name('profile.password.update');
+
+    Route::get('/{user_id}/orders}', [UserController::class, 'orders'])->name('profile.orders');
 
     Route::post('logout', [LoginController::class, 'logout'])
         ->name('logout');
@@ -56,5 +37,7 @@ Route::middleware('auth')->group(function () {
 Route::get('/products', [ProductController::class, 'index'])
     ->name('products.index');
 
-Route::get('/products/{slug}', [ProductController::class, 'show'])
+Route::get('/product/{slug}', [ProductController::class, 'show'])
     ->name('product.show');
+
+require __DIR__.'/auth.php';
